refactor: rename page-tracker
